### 2021-07-29 #3550 @whedon
[PRE REVIEW]: TraViA: a Traffic data Visualization and Annotation tool in Python
**Submitting author:** @OlgerSiebinga (<a href="http://orcid.org/0000-0002-5614-1262">Olger Siebinga</a>)
**Repository:** <a href="https://github.com/tud-hri/travia" target ="_blank">https://github.com/tud-hri/travia</a>
**Version:** v1.0.0
