### 2021-07-29 #3550 @kyleniemeyer
Hi @OlgerSiebinga, thanks for your interest in JOSS!

Visualization-only software does not always meet our requirements for research software be in scope (based [our definition](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software)), so I'm going to send this out to the editorial board for a scope check, before we move forward.
