### 2021-08-05 #3550 @OlgerSiebinga
Hi @danielskatz,
Good to hear that the paper has passed the scope review. As for potential reviewers, I looked at the list and found some people with domains/topics that seem relevant. I could not find any reviewers on the list that specified the exact topics of the paper, but here are my best results:

- MrMallIronmaker
- wiljnich; wiljnichepa
- 11michalis11
- rusu24edward
- nickbearman
- papachristoumarios
- bmoradi93
- arunmano121
- mponce0
- zeroset
- 1313e

Besides the potential reviewers on the list, another option would be one of the people that were involved in gathering the data (e.g. robertkrajewski one of the authors of HighD). But I don't know if they would be willing to review.
