### 2021-08-05 #3550 @danielskatz
👋 @bastilam, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews of open source software and short papers here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
