### 2021-08-09 #3550 @danielskatz
👋 @rusu24edward, @nickbearman, @papachristoumarios - Would one or two of you be willing to review this JOSS submission?
