### 2021-08-11 #3550 @danielskatz
👋 @tomalrussell - you've been recommended to me as a potential second reviewer for this submission by a new JOSS editor, @martinfleis.  We carry out our checklist-driven reviews of open source software and short papers here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html  Would you be willing to review this submission?
