### 2021-08-11 #3607 @whedon
[REVIEW]: TraViA: a Traffic data Visualization and Annotation tool in Python
**Submitting author:** @OlgerSiebinga (<a href="http://orcid.org/0000-0002-5614-1262">Olger Siebinga</a>)
**Repository:** <a href="https://github.com/tud-hri/travia" target ="_blank">https://github.com/tud-hri/travia</a>
**Version:** v1.1
**Archive:** <a href="https://doi.org/10.4121/16645651" target="_blank">10.4121/16645651</a>
