### 2021-08-11 #3607 @danielskatz
👋 @OlgerSiebinga - please note the DOI whedon suggests - if this is correct, please update your bib file then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
