### 2021-08-12 99ac529
removed unused entry from paper.bib
