+++ paper/paper.bib
@@ -97,9 +97,3 @@ title = {{PyKalman}},
 howpublished = {https://pykalman.github.io/},
 urldate = {01-04-2021}
 }
-@misc{PythonSoftwareFoundation,
-author = {{Python Software Foundation}},
-title = {pickle},
-howpublished = {https://docs.python.org/3/library/pickle.html},
-urldate = {2021-04-01}
-}
