### 2021-08-12 40f8aae
removed typos from the readme file
