+++ README.md
@@ -1,20 +1,21 @@
-# Travia: a TRAffic VIsualition and Annotation tool
+# Travia: a TRAffic VIsualization and Annotation tool
 
 In recent years many open datasets have been published that contain vehicle trajectories of human drivers on public roads. Such datasets can be useful for all
 kinds of research targeting traffic, e.g. traffic-flow studies and the development of autonomous vehicles. These datasets contain very large amounts of valuable
-data, however, they often focus on collecting data in one country and in one situation, e.g. highway traffic in Germany. To broaden the scope of a research
+data, however, they often focus on collecting data in one country and one situation, e.g. highway traffic in Germany. To broaden the scope of a research
 project, it can be valuable to include data from different datasets. This is often difficult because traffic data comes in *.csv files, often without a way to
-visualise and annotate the data for a quick start of your project. This is why Travia was created to combine te potential of different datasets by bringing them
-together in one tool for visualisation and annotation.
+visualize and annotate the data for a quick start of your project. This is why Travia was created to combine the potential of different datasets by bringing 
+them
+together in one tool for visualization and annotation.
 
-Datasets from three different projects can currently be visualized using Travia. The NGSIM project collected data using camera's on tall buildings next to roads
-in the USA. The PNeuma project captured all traffic during the morning rush-hour in the business district of Athens, for 5 days. To maximize precision and
+Datasets from three different projects can currently be visualized using Travia. The NGSIM project collected data using cameras on tall buildings next to roads
+in the USA. The PNeuma project captured all traffic during the morning rush hour in the business district of Athens, for 5 days. To maximize precision and
 coverage this was done using drones. Finally, the HighD project also used drones to record traffic on 6 locations at German highways.
 
-TraViA was build as a broad basis for visualisation and manual annotation, it can easily be extended to incorporate the specific needs of your research project.
+TraViA was built as a broad basis for visualization and manual annotation, it can easily be extended to incorporate the specific needs of your research project.
 Three examples of such specific implementations are included in Travia to illustrate the possibilities and provide a starting point for other developers. The
 first example considers automatically detecting and annotating specific situations in the HighD datasets. The second example is the plotting of heatmap
-overlays, to aid with the design of reward functions for autonomous driving. Finally it includes an example of plotting vehicle signals on an annotated
+overlays, to aid with the design of reward functions for autonomous driving. Finally, it includes an example of plotting vehicle signals on an annotated
 selection of data.
 
 Travia was tested with Python 3.8 on Windows.
@@ -48,7 +49,7 @@ download. This contains data recorded on the US-101 highway.
 ### Extract in the right folder
 
 To use the data, it has to be placed in the `travia/data` folder to be found automatically when loading the dataset. The data folder has three sub-folders, one
-for each data source. For our example we need the NGSim sub-folder, but if you have data from another project please see the `README.txt` file in the
+for each data source. For our example, we need the NGSim sub-folder, but if you have data from another project please see the `README.txt` file in the
 corresponding sub-folder for instructions. For now, navigate to your freshly downloaded data and extract the contents of `US-101-LosAngeles-CA.zip`
 to `travia/data/NGSim`. This should create a new folder, `travia/data/NGSim/US-101-LosAngeles-CA`, which contains more zip files. Extract the contents of
 `aerial-ortho-photos.zip` and `vehicle-trajectory-data.zip` to this folder. This will create more sub-folders, a representation of the correct end-result can be
@@ -57,9 +58,9 @@ found in the `README.txt` file located in `travia/data/NGSim`, please verify tha
 ### Alter `visualize.py` (run file)
 
 The script to start TraViA is a file called `visualize.py` and it can be found in the main project folder. This file is where you indicate which dataset you
[…]
