### 2021-08-12 tud-hri/travia#1 @tomalrussell
License information
Hi @OlgerSiebinga, I'm just starting a review for your JOSS paper (https://github.com/openjournals/joss-reviews/issues/3607).

Could you move or copy the GPL license in `COPYING` to `LICENSE.txt`, to check off one of the points on the review checklist?
