### 2021-08-13 tud-hri/travia#1 @OlgerSiebinga
Hi @tomalrussell, Thanks for taking the time to review my paper!

I thought a `COPYING` file was the standard for a GPL license, and also [acceptable for JOSS](https://joss.readthedocs.io/en/latest/review_criteria.html#software-license). But if you don't agree I'll change it, no problem.
