### 2021-08-13 tud-hri/travia#1 @tomalrussell
Ah no, that's fine - I'm more familiar with seeing `LICENSE` is all,  and this is the first time I'm reviewing for JOSS so wanted to dot the i's and cross the t's 🙂
