### 2021-08-21 #3607 @rusu24edward
@OlgerSiebinga, I've followed the setup instructions on the main Readme. When I try to run the visualize.py file, I get the following:
```
(v_travia) ~/codes/travia$ python3 visualize.py
QObject::moveToThread: Current thread (0x1cc8fb0) is not the object's thread (0x2165ab0).
Cannot move to target thread (0x1cc8fb0)

qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "/home/eddie/codes/travia/v_travia/lib/python3.8/site-packages/cv2/qt/plugins" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: xcb, eglfs, linuxfb, minimal, minimalegl, offscreen, vnc, wayland-egl, wayland, wayland-xcomposite-egl, wayland-xcomposite-glx, webgl.

Aborted (core dumped)
```
I'm using Python 3.8 on Ubuntu 18 and I've installed the requirements with pip as the Readme instructs.
