### 2021-08-25 tud-hri/travia#2 @tomalrussell
Travia visualize.py fails to start on Ubuntu
Following the recommended installation, also on Ubuntu with Python 3.8, in a clean environment, I see the same error that @rusu24edward reports in https://github.com/openjournals/joss-reviews/issues/3607#issuecomment-903069093

Is it possible that this is related to the error reported in this QT forum post on [using PyQT5 with OpenCV](https://forum.qt.io/topic/119109/using-pyqt5-with-opencv-python-cv2-causes-error-could-not-load-qt-platform-plugin-xcb-even-though-it-was-found)? Someone in that thread suggests switching to `opencv-python-headless`.

If I run:
```
pip uninstall opencv-python
pip install opencv-python-headless
python visualize.py
```
Then I get at least as far as the initial loading screen (processing data).

I suggest testing this on Windows too, then possibly updating `requirements.txt`
