### 2021-08-30 ce391ec
switched to opencv-python-headless because opencv-python gives loading errors when combined with pyqt un ubuntu
