### 2021-08-30 tud-hri/travia#2 @OlgerSiebinga
Tested this on Windows and it works. Updated the requirements.txt in commit ce391ec
