### 2021-08-30 decbc23
used os.path module for all file paths to ensure compatibility with other os's
