+++ processing/NGSIMsplitting.py
@@ -16,6 +16,8 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with Travia.  If not, see <https://www.gnu.org/licenses/>.
 """
+import os
+
 import pandas as pd
 
 from dataobjects.enums import NGSimDatasetID
@@ -29,8 +31,8 @@ def split_peachtree_data():
     seconds. This results in timestamps overlapping between the two datasets, so the datasets have to be manually separated and the timestamps have to be
     corrected for.
     """
-
-    all_peachtree_data = pd.read_csv('data/' + NGSimDatasetID.PEACHTREE_0400_0415.data_sub_folder + 'NGSIM_Peachtree_Vehicle_Trajectories.csv')
+    path_to_file = os.path.join('data', NGSimDatasetID.PEACHTREE_0400_0415.data_sub_folder, 'NGSIM_Peachtree_Vehicle_Trajectories.csv')
+    all_peachtree_data = pd.read_csv(path_to_file)
 
     # Find index of split, From a manual lookup the split between the datasets was found on the entry with Vehicle_ID 2 and Global_Time 1163030500
     split_index = all_peachtree_data.loc[(all_peachtree_data['Global_Time'] == 1163030500) & (all_peachtree_data['Vehicle_ID'] == 2), :].index
@@ -49,10 +51,11 @@ def split_peachtree_data():
 
         data_set['Global_Time'] = initial_timestamp * 1000 + milli_seconds_since_start
 
-    peachtree_0400_0415.to_csv('data/' + NGSimDatasetID.PEACHTREE_0400_0415.data_sub_folder + NGSimDatasetID.PEACHTREE_0400_0415.data_file_name + '.csv',
-                               index=False)
-    peachtree_1245_0100.to_csv('data/' + NGSimDatasetID.PEACHTREE_1245_0100.data_sub_folder + NGSimDatasetID.PEACHTREE_1245_0100.data_file_name + '.csv',
-                               index=False)
+    path_to_0400_0415 = os.path.join('data', NGSimDatasetID.PEACHTREE_0400_0415.data_sub_folder, NGSimDatasetID.PEACHTREE_0400_0415.data_file_name + '.csv')
+    path_to_1245_0100 = os.path.join('data', NGSimDatasetID.PEACHTREE_1245_0100.data_sub_folder, NGSimDatasetID.PEACHTREE_1245_0100.data_file_name + '.csv')
+
+    peachtree_0400_0415.to_csv(path_to_0400_0415, index=False)
+    peachtree_1245_0100.to_csv(path_to_1245_0100, index=False)
 
 
 def split_lankershim_data():
@@ -60,8 +63,8 @@ def split_lankershim_data():
     The lankershim data is collected in two sets, one between 08:28 - 08:45 am and the second between 08:45 - 09:00 am, both on june 15 2005.
     However the data is supplied in one data file. The timestamps in this file are correct, in contrary to the peachtree case.
     """
-
-    all_lankershim_data = pd.read_csv('data/' + NGSimDatasetID.LANKERSHIM_0845_0900.data_sub_folder + 'NGSIM__Lankershim_Vehicle_Trajectories.csv')
+    path_to_file = os.path.join('data', NGSimDatasetID.LANKERSHIM_0845_0900.data_sub_folder, 'NGSIM__Lankershim_Vehicle_Trajectories.csv')
+    all_lankershim_data = pd.read_csv(path_to_file)
 
     # Find index of split, From a manual lookup the split between the datasets was found on the entry with Vehicle_ID 2 and Global_Time 1118936700000
[…]
