### 2021-08-30 09bcc40
implemented GUI and argument parser to select dataset without altering the file
