+++ gui/__init__.py
@@ -1,2 +1,3 @@
 from .gui import TrafficVisualizerGui
 from .worldview import WorldView
+from .datasetselectiondialog import DatasetSelectionDialog
