+++ dataobjects/enums/datasource.py
@@ -23,4 +23,8 @@ class DataSource(enum.Enum):
     NGSIM = 0
     HIGHD = 1
     PNEUMA = 2
-    INTERACTION = 3
+
+    def __str__(self):
+        return{DataSource.NGSIM: 'NGSim',
+               DataSource.HIGHD: 'HighD',
+               DataSource.PNEUMA: 'pNEUMA',}[self]
