+++ gui/dataset_selection_dialog_ui.py
@@ -1,0 +1,64 @@
+# -*- coding: utf-8 -*-
+
+# Form implementation generated from reading ui file 'dataset_selection_dialog.ui'
+#
+# Created by: PyQt5 UI code generator 5.15.1
+#
+# WARNING: Any manual changes made to this file will be lost when pyuic5 is
+# run again.  Do not edit this file unless you know what you are doing.
+
+
+from PyQt5 import QtCore, QtGui, QtWidgets
+
+
+class Ui_DatasetSelectionDialog(object):
+    def setupUi(self, DatasetSelectionDialog):
+        DatasetSelectionDialog.setObjectName("DatasetSelectionDialog")
+        DatasetSelectionDialog.resize(320, 120)
+        self.verticalLayout = QtWidgets.QVBoxLayout(DatasetSelectionDialog)
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.label = QtWidgets.QLabel(DatasetSelectionDialog)
+        self.label.setObjectName("label")
+        self.verticalLayout.addWidget(self.label)
+        self.gridLayout = QtWidgets.QGridLayout()
+        self.gridLayout.setObjectName("gridLayout")
+        self.label_3 = QtWidgets.QLabel(DatasetSelectionDialog)
+        self.label_3.setObjectName("label_3")
+        self.gridLayout.addWidget(self.label_3, 1, 0, 1, 1)
+        self.label_2 = QtWidgets.QLabel(DatasetSelectionDialog)
+        self.label_2.setObjectName("label_2")
+        self.gridLayout.addWidget(self.label_2, 0, 0, 1, 1)
+        self.dataSourceComboBox = QtWidgets.QComboBox(DatasetSelectionDialog)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.dataSourceComboBox.sizePolicy().hasHeightForWidth())
+        self.dataSourceComboBox.setSizePolicy(sizePolicy)
+        self.dataSourceComboBox.setObjectName("dataSourceComboBox")
+        self.gridLayout.addWidget(self.dataSourceComboBox, 0, 1, 1, 1)
+        self.datasetComboBox = QtWidgets.QComboBox(DatasetSelectionDialog)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.datasetComboBox.sizePolicy().hasHeightForWidth())
+        self.datasetComboBox.setSizePolicy(sizePolicy)
+        self.datasetComboBox.setObjectName("datasetComboBox")
+        self.gridLayout.addWidget(self.datasetComboBox, 1, 1, 1, 1)
+        self.verticalLayout.addLayout(self.gridLayout)
+        self.buttonBox = QtWidgets.QDialogButtonBox(DatasetSelectionDialog)
+        self.buttonBox.setOrientation(QtCore.Qt.Horizontal)
+        self.buttonBox.setStandardButtons(QtWidgets.QDialogButtonBox.Cancel|QtWidgets.QDialogButtonBox.Ok)
+        self.buttonBox.setObjectName("buttonBox")
+        self.verticalLayout.addWidget(self.buttonBox)
+
+        self.retranslateUi(DatasetSelectionDialog)
[…]

[…]
