### 2021-09-08 #3607 @danielskatz
👋 @rusu24edward, @tomalrussell - can you let us know how you are doing, and what is currently blocking you from checking additional items?
