### 2021-09-09 #3607 @tomalrussell
Hi @OlgerSiebinga I've made a request for the HighD dataset but their one-time token download process hasn't worked for me and I haven't had any helpful response. Could you share a synthetic dataset which matches the HighD format so I can test the functionality which has been implement for that specific dataset?

Otherwise looking pretty good - testing is all manual and documented, rather than automated, but I read that as fine according to the review criteria.
