### 2021-09-10 #3607 @OlgerSiebinga
@tomalrussell, This was hard, but I managed to create some semi-realistic data in high format. Not all metrics are present, and something is messing up the timeline in the GUI for this data. But you should be able to verify the basic functionality with this dataset. Just load it as highd dataset 1. You can download the data [here](https://surfdrive.surf.nl/files/index.php/s/TrXnyUw1UCRFkYJ).
