### 2021-09-10 #3607 @danielskatz
@OlgerSiebinga - I wonder if you would want to make this data available as part of the documentation and testing for the software in general, not just for this review?  You could post the data in an archival repository such as zenodo.
