### 2021-09-10 #3607 @OlgerSiebinga
I think I'd rather keep the general testing the way it is. Currently, it is based on the publicly available NGSim dataset, which is real data and I think that's a better way to verify that the software is working than fake data. The only features that are specific for HighD data are all example implementations to illustrate what can be done when extending the software. I think those examples do not really show their full potential on this fake data. Besides that, researchers should be able to gain access to the real HighD data, it just takes a bit of time sometimes. So all-in-all I think it would be better to redirect users to the real datasets.
