### 2021-09-17 #3607 @tomalrussell
Yes - getting close. I now have the synthetic HighD dataset and still need to run through and understand the HighD-specific functionality. I expect to check this off early next week.

I'd also recommend making synthetic data available for all three datasets - it would then be possible for example to download them automatically and test the software without jumping through the non-open data hoops.

I'm not sure whether this is in scope for a review, but it's a related idea - the application could consider defining its own data format, that each example dataset is transformed into. This would be a nice step towards a metadata schema or specification for this kind of traffic data. It would help by documenting what the application expects as input. It could simplify the application code if some of the preprocessing is done outside of a more focussed GUI/visualiser. It would also suit constructing small examples to demonstrate features (or reproduce bugs).

I'd be happy to file these as suggestion-issues on the repository and not block on the review now, if that makes sense all round.
