### 2021-09-17 #3607 @danielskatz
great @tomalrussell - what you suggest seems good to me, with optional items created as issues that this review doesn't depend on
