### 2021-09-20 #3607 @danielskatz
@tomalrussell - given that you've checked off all the items now, are you satisfied to see this published?
