+++ paper/paper.md
@@ -21,7 +21,7 @@ bibliography: paper.bib
 
 # Summary
 
-In recent years, multiple datasets containing traffic recorded in the real world and containing human-driven trajectories were made available to researchers.
+In recent years, multiple datasets containing traffic recorded in the real world and containing human-driven trajectories have been made available to researchers.
 Among these datasets are the HighD, pNEUMA, and NGSIM datasets. TraViA, an open-source Traffic data Visualization and Annotation tool
 was created to provide a single environment for working with data from these three datasets. Combining the data in a single visualization tool enables
 researchers to easily study data from all sources. TraViA was designed in such a way that it can easily be extended to visualize data from other datasets and
@@ -29,22 +29,22 @@ that specific needs for research projects are easily implemented.
 
 # Statement of need
 
-The combination of drones, cameras, and image recognition techniques might sound like a recipe for a spy movie. But actually, this combination allowed for the
-collection of rich traffic datasets. The recipe is straightforward: hover a drone above a location with traffic, record a video and use image
+The combination of drones, cameras, and image recognition techniques might sound like a recipe for a spy movie. But actually, this combination allows for the
+collection of rich traffic datasets. The recipe is straightforward: hover a drone above a location with traffic, record a video, and use image
 recognition to generate bounding boxes for all vehicles. The result is a dataset containing human-driven trajectories at the location of interest that can be
-used for many scientific purposes, e.g. to study traffic flow, model human behavior, or design autonomous vehicle controllers.
+used for many scientific purposes, e.g., to study traffic flow, model human behavior, or design autonomous vehicle controllers.
 
-Because the required ingredients are easily accessed all over the world, multiple such datasets were published in recent years. In Germany, the highD
-project [@Krajewski2018] recorded all traffic on 6 different high-way locations, in Athens, Greece, all traffic in the city's business district was recorded
-using 10 drones for 5 days in the pNEUMA project [@Barmpounakis2020], and American highway traffic was recorded using fixed base cameras in the NGSIM
+Because the required ingredients are easily accessed all over the world, multiple such datasets have been published in recent years. In Germany, the highD
+project [@Krajewski2018] recorded all traffic at 6 different high-way locations; in Athens, Greece, all traffic in the city's business district was recorded
+using 10 drones for 5 days in the pNEUMA project [@Barmpounakis2020]; and American highway traffic was recorded using fixed base cameras in the NGSIM
 project [@NGSIM2016]. Combined, these datasets span different countries, types of vehicles, and environments, a combination valuable for researchers with
-different backgrounds. Example usages of these datasets are validating human behavior models (e.g. @Talebpour2015a; @Treiber2008) or testing autonomous vehicle controllers (e.g. @Schwarting2019). 
+different backgrounds. Example usages of these datasets are validating human behavior models (e.g., by @Talebpour2015a and @Treiber2008) or testing autonomous vehicle controllers (e.g., by @Schwarting2019). 
 
-Currently, it is difficult to leverage the powerful combination of multiple datasets because all datasets come in different formats, and it is often
-difficult to get a good and real-time visualization of the data. Some visualization tools exist (one is provided with the highD data [@Krajewski2018] another 
+Currently, it is difficult to leverage the powerful combination of multiple datasets because all the datasets come in different formats, and it is often
+difficult to get a good and real-time visualization of the data. Some visualization tools exist (one is provided with the highD data [@Krajewski2018] and another 
 example for NGSIM data can be found in @Sazara2017) but they are specifically made for only one of these datasets and are very basic in the sense that they 
-provide little control over simulation time and no insight in raw values per vehicle per frame. Besides difficulties with
[…]
