### 2021-09-20 #3607 @danielskatz
👋 @OlgerSiebinga - I've suggested some minor changes to the paper in https://github.com/tud-hri/travia/pull/9

At this point, could you:
