### 2021-09-20 #3607 @OlgerSiebinga
I've done all the mentioned to-do's. A tagged release was made with version tag v1.1. The software is uploaded to the 4TU research data repository. It still has to be reviewed there before it becomes public. The archive DOI is 10.4121/16645651
